Use XUnit source generator for regression tests Split tests into separate projects
