+++ Tests/Regression/Bug_206/Bug_206.csproj
@@ -19,12 +19,20 @@
 	  <UseAppHost>false</UseAppHost>
 
 	  <OutputType>Exe</OutputType>
+
+    <EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
+    <CompilerGeneratedFilesOutputPath>.\GeneratedFiles</CompilerGeneratedFilesOutputPath>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\System.Private.CoreLib\System.Private.CoreLib.csproj" />
     <ProjectReference Include="..\..\..\ILCompiler\ILCompiler.csproj" />
     <ProjectReference Include="..\..\Common\Tests.Common.csproj" />
+    <ProjectReference Include="..\..\XUnit.SourceGenerator\XUnit.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
   </ItemGroup>
-  
+
+  <ItemGroup>
+    <Compile Remove="$(CompilerGeneratedFilesOutputPath)/**/*.cs" />
+  </ItemGroup>
+
 </Project>
