+++ Tests/Regression/Bug_468/SpillImportAppendTests.cs
@@ -1,7 +1,15 @@
-﻿namespace Regression
+﻿using Xunit;
+
+namespace Regression
 {
     internal class SpillImportAppendTests
     {
+        [Fact]
+        public static void Test()
+        {
+            Assert.Equal(2, new SpillImportAppendTests().SpillOnStFldImport());
+        }
+
         private int x = 2;
         public int SpillOnStFldImport()
         {
