+++ Tests/Regression/Bug_623/Bug_623.csproj
