+++ Tests/Regression/Bug_623/bug623.il
