+++ Tests/Regression/Bug_655/SpillImportStElemTests.cs
@@ -1,8 +1,16 @@
-﻿namespace Regression
+﻿using Xunit;
+
+namespace Regression
 {
     internal static class SpillImportStElemTests
     {
-        public static int Test()
+        [Fact]
+        public static void Test()
+        {
+            Assert.Equal(37, StElemShouldSpillWhenAppendingStatement());
+        }
+
+        public static int StElemShouldSpillWhenAppendingStatement()
         {
             object?[] arr = new object[1];
             arr[0] = 37;
