+++ Tests/Regression/RegressionTestsRunner/RegressionTests.csproj
@@ -19,7 +19,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.0" />
     <PackageReference Include="NUnit" Version="4.2.1" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.6.0" />
+    <PackageReference Include="NUnit3TestAdapter" Version="4.6.0" />    
   </ItemGroup>
 
   <ItemGroup>
